requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
cssselect2==0.8.0
Django==4.2.16
django-crontab==0.7.1
djangorestframework==3.16.0
fonttools==4.58.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
reportlab==4.4.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-extensions==4.12.2
tzdata==2024.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
